fixup! Refactor NameDefinition to use ADTs and reuse code
