+++ modules/spark/src/main/scala/notebook/kernel/repl/common/ReplShared.scala
@@ -1,7 +1,5 @@
 package notebook.kernel.repl.common
 
-import org.apache.spark.repl.SparkIMain
-
 abstract class NameDefinition {
   def name: String
   def tpe: String
@@ -12,22 +10,3 @@ object NameDefinition {
 }
 case class TypeDefinition(name: String, tpe: String, references: List[String]) extends NameDefinition
 case class TermDefinition(name: String, tpe: String, references: List[String]) extends NameDefinition
-
-object SparkInterpUtils {
-  implicit class SparkInterpreterExtensions(interp: SparkIMain) {
-    def getTypeNameOfTerm(termName: String): Option[String] = {
-      val tpe = try {
-        interp.typeOfTerm(termName).toString
-      } catch {
-        case exc: RuntimeException => println("Unable to get symbol type", exc); "<notype>"
-      }
-      tpe match {
-        case "<notype>" => // "<notype>" can be also returned by typeOfTerm
-          interp.classOfTerm(termName).map(_.getName)
-        case _ =>
-          // remove some crap
-          Some(tpe.replace("iwC$", ""))
-      }
-    }
-  }
-}
